moved list_to_str function to string_manip util and moved tests to appropriate area
